change behavior of `rag rm`
Previously, `rag rm **/*.py --staged` expected all the python files to
be staged. If there's a processed python file, it threw an error and did
nothing.

Now, if there are staged python files and processed python files, `rag
rm **/*.py --staged` only removes staged ones and ignores the processed
ones. It still throws an error if the argument includes an untracked
file.

+++ src/main.rs
@@ -1378,6 +1378,10 @@ async fn run(args: Vec<String>) -> Result<(), Error> {
                 )?;
             }
 
+            if result.is_empty() {
+                return Err(Error::NoFileToRemove);
+            }
+
             if !dry_run {
                 for file in files.iter() {
                     index.remove_file(
